Fixes regex issue when selecting with multiple attributes
  * Adds more unit tests for set and rm
  * Updates metadata file to use voxpupuli namespace
  * Moves dummy lens provider to puppet_x utility namespace
