+++ Gemfile
@@ -30,6 +30,7 @@ group :test do
   gem 'coveralls',                                                  :require => false
   gem 'simplecov-console',                                          :require => false
   gem 'parallel_tests',                                             :require => false
+  gem 'compare-xml'
 end
 
 group :development do
