+++ spec/unit/puppet/provider/xmlfile/posix_spec.rb
