+++ spec/unit/puppet/provider/xmlfile/windows_spec.rb
